:truck: rename components
