+++ src/pages/home/HomePage.tsx
@@ -1,10 +1,10 @@
 import { useQuery } from '@tanstack/react-query';
 import { useTranslation } from 'react-i18next';
 
+import CategoryTitle from '@components/bar/CategoryTitle';
+import Title from '@components/bar/Title';
+import BookCarousel from '@components/carousel/BookCarousel';
 import UserHeader from '@components/header/UserHeader';
-import CategoryTitle from '@components/layout/CategoryTitle';
-import Title from '@components/layout/Title';
-import BookSlider from '@components/slider/BookSlider';
 import { Book, BookCardModel } from '@models/book';
 import { BookService } from '@services/BookService';
 
@@ -88,10 +88,10 @@ const HomePage = () => {
           rightLabel="더보기 >"
           onClickRightLabel={() => alert('imhere')}
         />
-        <BookSlider books={sampleBooks} />
+        <BookCarousel books={sampleBooks} />
 
         <CategoryTitle category="컴퓨터" />
-        <BookSlider books={sampleBooks} />
+        <BookCarousel books={sampleBooks} />
       </>
     );
   }
